Add files via upload
In this dataset, there are 90 rows and 31 columns.

11 columns are of float datatype, 3 columns of int datatype and 17 columns of object datatype.

13 columns contain null values.

There is no duplicate row present.

10 unique genres of movies are present in this dataset.

**Lead Actor 1 -->** Average rating is 6.12. Rahul Bhat and Satyajeet Dubey got the highest rating of 8.8 and Riteish Deshmukh got the least rating of 2.

**Lead Actress -->** Average rating is 6.1. Genelia D'Souza got the lowest rating of 2 and Simran,Priya Bapat and Revathi got the highest rating of 8.8.

**Lead Actor 2 -->** Average rating is 6.17. Suriya, Alka Amin and Mrinmayee Godbole got the highest rating of 8.8 and Mahesh Manjrekar got the least rating of 2.

**Director -->** Average rating is 6.13. R. Madhavan, Satish Munda and Anirban Bose got the highest rating of 8.8 and Shaad Ali got the least rating of 2.

**Production House -->** Average rating is 6.12. Tricolour Films, Zee Music Company and Platoon One Films Pvt. Ltd. got the highest rating of 8.8 and Soundarya Productions got the least rating of 3.5.

**Distribution Company -->** Average rating is 6.04. UFO Moviez got the highest rating of 8.8 and Soundarya Productions got the least rating of 3.5

**Music By -->** Average rating is 6.15. Meghdeep Bose got the highest rating of 8.8 and Sneha Khanwalkar got the least rating of 2.

Average number of screens in India is 1830 screens. Pathaan movie got the highest number of screens(5500) and Rocket Gang movie got the least number of screens(300).

Average number of views on trailer on Youtube is 28661777 views. Trailer of Sooryavanshi movie got the highest views and Trailer of Trahimam movie got the lowest views.

Average Marketing Expense is 96062500 or 9.6 crores. Highest amount was spent in the marketing of Raksha Bandhan movie and Lowest amount was spent on the marketing of Dedh Lakh Ka Dulha movie.

Average Budget is 628915662.65 or 62.89 crores.

Highest budget movie in this dataset is Brahmāstra: Part One – Shiva and Lowest budget movie in this dataset is Bhavai.





Some of the columns are highly correlated with other.

There is positive linear relationship between some of the variables.

10 columns have outliers.

6 columns are left tailed or negatively skewed.

6 columns are right tailed or positively skewed.